requirements.txt
Flask==2.3.3
requests==2.31.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
cachelib==0.10.2